Make list tokenizer, linked lists are up next

+++ Output/dad_example.cpp
@@ -1,6 +1,6 @@
 #include "../Std_Lib/std_lib.hpp"
-auto x = mul(3, 8, 5, (int)(add(2, 2, 3)));
+auto x = mul(3, 8, 5, add(2, 2, 3));
 int main() {
-display(x);
-return 0;
-}
+	display(x);
+	return 0;
+	}
